pyproject.toml
[tool.poetry]
name = "c2modregistry"
version = "0.1.0"
description = ""
authors = ["Jacob Barber <jacoby6000@gmail.com>"]
readme = "README.md"
packages = [{include = "c2modregistry", from="src/"}]

[tool.poetry.dependencies]
python = "^3.10"
PyGithub = "^1.59.1"
requests = "^2.31.0"
types-requests = "^2.31.0.2"
argparse = "^1.4.0"
natsort = "^8.4.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
